Smoke sensor
Use description field in smoke sensor to distuingish it from on On/Off
switch. This can be used for smoke sensors which only have On/Off switch
charecteristics in Domoticz, such as a Fibaro FGSD002.

If you add "Reacticz: smoke sensor" to the description field in
Domoticz,  it will show as a smoke sensor on the Reacticz dashboard in
stead of an On/Off switch.